Drop PSR SimpleCache
The way that interface works with multiple versions has in practice
become a nightmare of dependency management.

V1, V2 and V3 of simple cache are not compatible with each other yet
use the same name and same namespace. So that you can't have a project
using 2 libraries when one expects and requires V1 and the other expects
and requires another version.

Had those interfaces been named CacheInterface, CacheInterfaceV2 and
CacheInterfaceV3 then it would have been possible to have multiple
versions, and even type hints like `CacheInterface | CacheInterfaceV2`
would've worked.

Currently none of this is possible and it's a mess.